requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
psutil>=5.9.0
docker>=6.1.0
